feat: menu bar component
