+++ src/pages/HomePage.js
@@ -1,18 +1,14 @@
 import React, { useState, useEffect } from "react";
-import { FaSearch, FaBars, FaTimes, FaTrash } from "react-icons/fa";
+import { FaSearch, FaTrash } from "react-icons/fa";
 import styles from "../css/HomePage.module.css";
 import Header from "../components/Header.js";
 import Footer from "../components/Footer.js";
 import Modal from "../components/Modal.js";
 import Success from "../components/Success.js";
 import SideBar from "../components/SideBar.js";
+import MenuBar from "../components/MenuBar.js";
 
 function HomePage() {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const toggleMenu = () => {
-    setIsMenuOpen(!isMenuOpen);
-  };
-
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [currentEvent, setCurrentEvent] = useState(null);
 
@@ -135,36 +131,7 @@ function HomePage() {
             </ul>
           </section>
         </main>
-        <button
-          className={` ${styles.button} button absolute top-4 right-4 text-blue-500 p-2 rounded`}
-          onClick={toggleMenu}
-        >
-          {isMenuOpen ? <FaTimes /> : <FaBars />}
-        </button>
-        <aside
-          className={`w-64 bg-gray-100 p-4 flex flex-col absolute right-0 top-0 bottom-0 transition-transform transform ${
-            isMenuOpen ? "-translate-x-0" : "translate-x-full"
-          }`}
-        >
-          <h2 className="text-2xl font-bold mb-4">Menu</h2>
-          <ul>
-            <li className="mb-2">
-              <a href="#" className="text-blue-500">
-                Profile
-              </a>
-            </li>
-            <li className="mb-2">
-              <a href="#" className="text-blue-500">
-                Settings
-              </a>
-            </li>
-            <li className="mb-2">
-              <a href="#" className="text-blue-500">
-                View Legacy Data
-              </a>
-            </li>
-          </ul>
-        </aside>
+        <MenuBar></MenuBar>
       </div>
       <Footer />
     </div>
