+++ src/pages/InjuryPage.js
@@ -1,7 +1,13 @@
 import React from "react";
 
 function InjuryPage() {
-    
+  return (
+    <div className="max-h-screen flex h-screen flex-col">
+      <Header path="/home"></Header>
+      <MenuBar></MenuBar>
+      <Footer></Footer>
+    </div>
+  );
 }
 
 export default InjuryPage;
